Update communication module to handle failures to connect to IBRDTN
It handles the following scenarios:
1 - Sensor Node initialization:
    At initialization, sensor node will create communication module,
when communication module is created it attempts to connect to IBRDTN daemon.
If IBRDTN daemon has NOT finished its initialization, it will not be possible to
connect to its API, then communication module will throw an error. To handle this
case, communication module tries to connect a number of times, with a fixed time interval
between each try. If it fails in all tries to connect, then you have two options:
    I - IBRDTN has a lot of data/bundles to load, then its necessary wait a bit more
before attempt a connection. In this case you either increase the time between connection
tries or increase the ammount of tries.
    II - IBRDTN has some bug during initialization: ¯\_(ツ)_/¯ cant do anything here

2 - Socket disconnects from IBRDTN before trying to send a message: In this case,
it will try to reconnect to IBRDTN daemon as mentioned above, if it reconnects, then
the message will be tried to be sent. But if it fails to reconnect then sensor node
will be restarted.